bg_atlasapi -> brainglobe_atlasapi, package restructure
