+++ .gitignore
@@ -1,4 +1,4 @@
-bg_atlasapi/bg_config.conf
+brainglobe_atlasapi/bg_config.conf
 
 workspace.py
 .vscode/
