+++ brainglobe_atlasapi/descriptors.py
